Bug#36104070 max-connection-error reached on connect/disconnect
When a external load-balancer does a health check against
the router-ports, it usually opens a TCP connection, checks it doesn't
fail and closes the connection again.

Currently, this open+close is treated as "max_connect_error" which
may lead to the router closing the incoming port as the
max-connect-errors limit is reached (if no valid connection is done
on that port).

Change
======

- Do not treat open+close as "connect-error" anymore.
- Only protocol-errors inside the handshake are now treated as
  connect-error.

Change-Id: I011bfd99c480a6375557340579b9ea09ff952157