Bug#35444004: Result diff seen with hypergraph off and on (different number of rows)
Some implicitly grouped queries using the MIN or MAX aggregation
functions could return wrong results if the MIN or MAX value was
fetched from an index and the WHERE clause contained a BETWEEN
predicate.

The bug was caused by incorrect logic for combining predicates in the
WHERE clause. For example, x <= 1 AND x BETWEEN 3 AND 5 was reduced to
x BETWEEN 3 AND 5. Since the former rejects all rows and the latter
accepts some, wrong results could be returned. Another problem was
with predicates like x < 5 AND x BETWEEN 2 AND 4, which was correctly
reduced to x BETWEEN 2 AND 4, but where the index read was set up to
read only values less than 4, so rows with x = 4 were not seen even
though they should have been.

The bug does not affect the hypergraph optimizer, because it does not
use the code path which contains the bug.

Two fixes were needed:

1) When determining if a BETWEEN predicate can help lower the upper
bound of a MAX scan, compare the current upper bound to the "max"
operand of the BETWEEN predicate, and ignore the "min" operand, as it
should not have any impact on the upper bound of the scan. Similarly,
for MIN scans, only use the "min" operand of the BETWEEN predicate to
see if the lower bound can be increased.

2) When using a BETWEEN predicate to set the upper or lower bound of
the index scan, update the range flags to reflect that it's a closed
interval, so that the flags set when processing a previous < or >
predicate aren't kept and make the scan lose rows at the boundary of
the range.

Change-Id: Ide0ceb0302729b3322b112cdf2768596c290c7ab