Bug#36105279 access denied for nodejs/mysql
If Router is running with
client_ssl_mode=DISABLED/server_ssl_mode=DISABLED and nodejs/mysql is
used, authentication fails through Router, while the same works when the
mysql client is used.

nodejs/mysql does not support the CLIENT_PLUGIN_AUTH capability and only
supports mysql_native_password.

If CLIENT_PLUGIN_AUTH is enabled, the Router can force a
"auth-switch-method" between server and client and send the server's
nonce to the client and make the auth succeed. When CLIENT_PLUGIN_AUTH
is not supported by the client, that's not possible and auth will always
fail.

Change
======

- Fail the handshake with clients that don't support CLIENT_PLUGIN_AUTH
  if the router sends the initial greeting (not PASSTHROUGH, nor
  PREFERRED/AS_CLIENT)

Change-Id: I7ae2bbb7bb6fe5d34de3b58d5bab771b3bb40094