Bug#34057432: MAX gives a different answer after an index is added
Some implicitly grouped queries using MAX or MIN on a string column
could return wrong results if there was an index on the string column
and a predicate in the WHERE clause compared the column to a numeric
constant.

The reason for the wrong results was that the optimizer decided that
the MAX or MIN values should be read from the index, and by doing
this, it changed the semantics of the predicate in the WHERE clause.
Comparisons between strings and numbers should be performed by
implicitly converting the arguments to floating point numbers before
comparing them. When using the index, it instead converted the numeric
constant to string and compared the values as strings.

Fixed by disabling this particular optimization when a predicate in
the WHERE clause performs comparison with different semantics than an
index lookup.

Change-Id: I2555f9d8f505b71c950e0ccdd9a44e8f09b7f905