Bug#36179509 NdbInfo restart_info reports wrong state when there are no subscribers
Problem:
When a Node is started with no nodegroup assigned and
there is no subscribers connected in the cluster the
NdbInfo restart_info will be reported as "Wait handover
of subscriptions" instead the expected state "Restart
completed".
This happens because when there are no buckets to handover
nor subscriptions reports to send out dict lock is not
locked and this way the update of the ndbinfo.restart_state
is not done. Update of the ndbinfo.restart_state is triggered
by the dict unlock.

Solution:
In check_start_handover force dict lock also when there
are no buckets to handover or subscriptions reports to
send out. This way DIH will change restart_state of the
starting node to "Restart completed".

Change-Id: I07d3ef6f4e28adf65dd59fad74fe6947f70c5990