Bug#30112096: ASSERTION `HAVING_COND->HAS_SUBQUERY() ||               !(HAVING_COND->USED_TABLES() & ~(1 | (
When adding a having condition to temp table, it is expected that all
the fields in the having condition are already replaced with the
temp table fields. However, for a query which has expression involving
"Item_row" type in the HAVING clause, constant expressions are not
getting cached(finalize_table_conditions()) resulting in HAVING clause
still having references to the fields from the underlying tables. And
the reason for this is the missing implementation of ::compile()
for the Item_row class which actually does the replacements with
constants.

Solution is to add ::compile for Item_row class.

Change-Id: I6e4a4b2fe79e096093d5b5262edf9256dd2440f5