Bug#36049906 Cluster/J Bulk Deletes
It is a documented limitation of Cluster/J Query that setLimits() cannot
be used with deletePersistentAll(). This patch removes that limitation,
in order to allow a user to perform a large delete in batches, using a
batch size that is smaller than MaxNoOfConcurrentOperations.

This change allows use of a limit size, but not a skip count,
in setLimits(). Use of a non-zero skip count will cause
deletePersistentAll() to fail.

To make use of this feature, the user should delete iteratively until
the count of deleted rows is less than the batch size, as in this
example:

    /* Delete in batches */
    query.setLimits(0, DeleteBatchSize);
    int result = 0;
    do {
        result = query.deletePersistentAll();
        System.out.println("Batch result: " + result);
    } while(result == DeleteBatchSize);

Change-Id: I7b7337ce75e4f7529028dccb5455bd1fcaa3051a