Bug#35105475: Result diff seen with hypergraph off and on (different number of rows)
Wrong results were seen in some queries that were comparing string
columns with numeric columns. The bug affected both the old optimizer
and the hypergraph optimizer, but not necessarily in the same queries,
so the results could vary depending on which optimizer was used.

The wrong results were caused by equality propagation being too eager
and rewriting expressions in ways that changed their semantics. For
example, it could rewrite a WHERE clause such as ...

  WHERE varchar_col1 = varchar_col2 AND varchar_col1 < int_col

... to ...

  WHERE varchar_col1 = varchar_col2 AND varchar_col2 < int_col

Because the equality predicate uses string semantics and the less than
predicate uses floating-point semantics, this transformation is not
valid. If varchar_col1 has the value '4' and varchar_col2 has the
value '᮴' (which is SUNDANESE DIGIT FOUR), the equality predicate will
consider them equal (assuming the default utf8mb4_0900_ai_ci collation
is used). But when compared to an integer column, the string value is
converted to a floating-point number, and in numeric context they are
not considered equal. The former will be implicitly converted to 4.0,
whereas the latter will be implicitly converted to 0.0. And the
expression 4.0 < int_col is different from 0.0 < int_col.

The fix is to disable equality propagation if the field reference is
used in a different comparison context than that of the replacement.
This can be checked with Item::has_compatible_context(). It is already
used in constant propagation, but not in equality propagation.

In the hypergraph optimizer, a check for
Item_func::allow_replacement() is added in addition to
Item::has_compatible_context(), since it was forgotten when
Bug#35049440 was fixed. The test case for Bug#35049440 is also
extended with a case that shows the bug was not fixed in the
hypergraph optimizer.

Change-Id: I2e6e95b80e2a08fb8fdc8c7bbbc1f5fa7167baed