Bug#36112585: Data type is changed from VAR_STRING to STRING by UNION if all values are empty string
When selecting two empty strings that are combined with UNION as in:

  SELECT '' AS a UNION SELECT '';

the resulting datatype is CHAR(0) instead of VARCHAR(0). This is
contradictory to the documentation (see function CASE) which says:

  If all types are character string (CHAR or VARCHAR), the result is
  VARCHAR with maximum length determined by the longest character
  length of the operands.

The fix is to remove the exception for strings with length 0.

A few test result files have been updated with minor metadata changes
from BINARY(0) to VARBINARY(0) or CHAR(0) to VARCHAR(0).

Change-Id: I6a86461b98fd527ee49b84e15b12d08c92600f7f