Bug#36132430 ndb_waiter has scrambled output when finding tls certificates
When running ndb_waiter with a path with certificates it provides
scrambled output due to a missing new line:

  $ ndb_waiter --ndb-tls-search-path=mysql-test/std_data/ndb-tls/active
  Using Certificate: mysql-test/std_data/ndb-tls/active/ndb-api-certConnecting to management server at nodeid=0,localhost:1186

This change suppresses the certificate path output and appends an
indication to the connecting message whether it tries to use tls or not.

  $ ndb_waiter --ndb-tls-search-path=mysql-test/std_data/ndb-tls/active
  Connecting to management server at nodeid=0,localhost:1186 (trying TLS)

To test with mtr add a tls.cnf file which sets ndb-tls-search-path and
run ./mtr --defaults-extra-file=tls.cnf ndb.ndb_tools_connect.

Change-Id: I2b4060f223a049bcbb77b6aa0bab42c1cccf8ba0