Bug#36142806 	innodb_parallel_read_threads>1 makes simple select count more expensive
Symptom:
A full table scan is much slower with `innodb_parallel_read_threads` of 2 than with value of 1 when IO reads from disk are issued to fulfill the scan. The higher values of `innodb_parallel_read_threads` show any gain in processing time only if the table is already fully cached in the Buffer Pool.

Root cause:
With `innodb_parallel_read_threads` value of 1 during scan a read mode of `Page_fetch::NORMAL` is used. It causes the Buffer Pool blocks to get `access_time` assigned and the linear read-ahead to be triggered.
With value of 2 or more, the `Page_fetch::SCAN` is used, which causes the `access_time` to not be set, which makes the `buf_read_ahead_linear()` to see the page as "not accessed" and returns without doing the read-ahead. Additionally, there is an if condition that is explicitly disabling a call to `buf_read_ahead_linear` for this fetch mode, but it looks like it was an optimization to not call the `buf_read_ahead_linear()` as it would not do anything.

Fix:
The `Buf_fetch<T>::single_page()` is setting `access_time` for block when `Page_fetch::SCAN` is used and removes the `if` condition that disabled the call to the `buf_read_ahead_linear()`.
This makes the read ahead to work as intended.

Change-Id: Ifec6b80b1bee3240131bf733f229729c3cf3afc8