Bug#36118112 	[InnoDB] srv_io_thread_function are assigned in incorrect order
Symptom:
The names of threads in `SHOW ENGINE INNODB STATUS` in the `FILE I/O` section are wrong - the first name is `(null)`, rest is shifted by one.

Root cause:
A shift by one was introduced by Bug#34992157.

Fix:
The names are filled in in `AIO::start()` in correct order.

Change-Id: I6e752c75d5778ae524f4494574891e9ec8040739