Bug#36135001 Hypergraph: Too low row estimate for index lookup [2/2]
This commit is the actual bugfix. It adds checks to ensure that we do
not add two equivalent conditions to m_pushable_conditions of the
same JoinHypergraph::Node instance. This prevents the selectivity
of the same predicate to be applied twice (which will e.g. give
a total selectivity of 1e-6 instead of 1e-3) for index lookups.

Change-Id: I3bfd4af053fe84e194137083344178d62a16bb63