WL#15924 Support upgrade from 8.x only
The aim of this WL is to allow upgrade only from 8.x.
An attempt to upgrade from 5.7 will be disabled. Upgrade from
previous 8.0 release which had data dictionary will be allowed.

This allows us to remove lot for code in server and InnoDB that
deals with 5.7 specific logic, e.g., code that deals with .frm
file format. This WL will remove some but not all source code
that is used to support upgrade from 5.7. Further removal will
have to be done in separate worklogs at a later stage.

Following are changes made:
* Throw error ER_DD_SE_INIT_FAILED when upgrading from 5.7 database.
* Removed dependency of start-up code with 5.7 upgrade call flow.
* Removed some parts of server code that is deals with 5.7 upgrade
  during server startup. This removes two instances of bootstrap
  thread creation to handle 5.7 metadata migration.
* Removed some parts of InnoDB code related to DICT_INIT_UPGRADE_57_FILES.
* Removed Table_trigger_dispatcher::reorder_57_list.
* Removed files in sql/dd/upgrade_57.
* Removed all the *.zip files containing 5.7 databases.
* Removed related tests that use 5.7 databases.

Change-Id: I049a81cabc0c19353a7200d968abc5e47adfcbe3