Bug#36032958: Assertion `IsEmpty(child.delayed_predicates)' failed(HYPERGRAPH=ON) [2/2, fix]
An assertion failed in the hypergraph optimizer for a query that used
an index group skip scan for GROUP BY, and had a window function, and
the WHERE clause contained a non-deterministic predicate.

The assertion in itself was harmless. It was supposed to remind us to
copy the delayed_predicates bit set from the child access path and
into the STREAM access path if the code changes so that STREAM access
paths can be added before all filters have been applied. The
particular STREAM access path that hit the assertion was added between
aggregation and windowing, and all the delayed predicates should have
been applied well before that.

The problem was that ApplyPredicatesForBaseTable() had incorrectly
marked a predicate as a delayed predicate. It was a non-deterministic
predicate which referenced the base table. Because of the
RAND_TABLE_BIT, it thought the predicate referenced the base table and
one other table, and could be applied as a delayed predicate once that
other table was joined in. But there is no table corresponding to
RAND_TABLE_BIT, and anyway ApplyDelayedPredicatesAfterJoin() won't
apply any non-deterministic predicates, so it was never applied as a
delayed predicate, and the bit for this predicate was never removed
from the delayed_predicates bit set.

The predicate was eventually applied, but not as a "delayed predicate"
while creating the join paths, but rather as a "final predicate" which
gets applied after all the tables have been joined together. All
non-deterministic predicates, as well as predicates not referencing
any tables, are applied as final predicates after all tables have been
joined, regardless of whether they are marked as delayed or not. But
the assertion in EstimateStreamCost() doesn't know that, and it gets
confused when it sees that the bit is set.

The fix is to make ApplyPredicatesForBaseTable() not mark
non-deterministic predicates as delayed predicates. The predicate gets
applied anyway, and when the bit is not set, the assertion in
EstimateStreamCost() doesn't fail.

For good measure, a new assertion is added right before the final
predicates are applied. It checks that all the delayed predicates have
been applied when we get there.

Change-Id: Ief039e76ffed3cf6d99ac9b6ffafd80fdbca7fe4