Bug#35560806 Aggregate on window function SEGV [ back-port ]
[ Change relative to trunk version pf the patch: remove QUALIFY tests:
feature not in available until 8.3 ]

Missing check during resolution allowed an ORDER BY containing a
grouped aggregate function with an argument being a SELECT list alias
of a window function, causing the issue: window functions are
performed after grouping, so trying to invoke a window function inside
a grouped aggregate function breaks that order and leads to the SEGV.

Added missing check.

This patch also solves the duplicate case of Bug#36022337 "Order by
with window function SEGV". The tests include both repros, i.e. the
one for this issues and the one for Bug#36022337.

It also solves duplicate Bug#35845444 cases: window alias inside
grouped aggregate in HAVING or QUALIFY predicates.

Change-Id: I6935557c92d59f0045b15a8a218d30be30333289