WL#16087: Remove the usage of the deprecated SHOW SLAVE STATUS command in tests and other components [Step 2]
This worklog removes the usage of SHOW SLAVE STATUS in tests files and
other minor usages. It also renames most of the usage of some of its
columns, like Slave_IO_State or Exec_Master_Log_Pos.

Step 2: Remove the usage of the command and some of its columns on server code

Change-Id: I4a7945690d8a67da26621db95f57f095a932668c