Bug#36125903 Hypergraph: Inconsistent numeric precision for AccessPath in optimizer trace
When using the Hypergraph optimizer and enabling optimizer trace, the
trace would (prior to this commit)  describe candidate access paths
like this:

{HASH_JOIN, cost=70202.0, init_cost=101.0, rescan_cost=70102.9,
rows=1000000.0, join_order=(x1,x2)}

Cost figures and row estimates were presented as decimal numbers with at
most one digit after the decimal point. This had two unfortunate
consequences:

* Any number below 0.05 was presented as zero.

* Large numbers (e.g. 10e9) would have lots of digits and therefore
  be hard to read.

This commit ensures that these figures are presented in the same format
as in "EXPLAIN" output. Numbers in the [0.001, 999999.5> range are
presented as decimal numbers, and other numbers are presented in
scientific format (see also fix for Bug#34114563 for details).

Change-Id: I44a04c7ec78e494f7fb988e5b3e32104af3844a1