Bug#34901428 [noclose] Modernize JTie: Errant return type
A compiler warning in jtie_tconv_array_impl, "variable 's' set
but not used", revealed a bug in the template code.

The existing return statement caused C *c to be implicitly
cast to a cstatus and then returned. Return s, as documented.

Note that two functions ObjectArrayConvImpl::copyToCObjectArray()
and ObjectArrayConvImpl::copyToJavaObjectArray(), affected by this
change, are actually unused. Martin Zaun noted this in his comment
at line 280: "while assumed to work, this mapping has not been used
and tested yet."

Change-Id: Ib1a4a3cb28bfac9f3ea0506b2da036e3c642ec6b