BUG#36054355: Set network namespace errors not reported on replication receiver status
When a replication channel is configured with a unknown network
namespace, through the parameter `NETWORK_NAMESPACE` of the `CHANGE
REPLICATION SOURCE` command, on channel start the channel receiver
thread is stopped but the unknown network namespace error is not
recorded on the channel internal status neither reported on
`performance_schema.replication_connection_status` table.

This patch adds the missing error handling and reporting.

Change-Id: I6a7111cbff5e6f1a39afe5eaa2952bc885b24c1c