Bug#36159678 (4 of 6) JTie: Fix some ndbjtie function mappings
The jtie mapping for NdbOperation::getNdbErrorLine() should be marked
as const-overloaded.

The jtie mapping for Ndb_cluster_connection::set_service_uri() was
declared as void, but should actually return int.

The mappings for three const methods were defined to use the _t
variant of the object trait type, but need to use the _ct variant.

Change-Id: I4ab9e7a093c74edef72e2141d701e2e6c32b8bb6