WL#15699 Reject foreign key references to non unique key and partial key
Disallow the creation of foreign key that reference non-unique key
and partial fields of composite keys which is non standard behavior

1. Added restrict_fk_on_non_standard_key session variable. This
variable is deprecated and will be removed in future release.
2. Implemented function check_non_standard_key_exists_in_fk() which
checks whether referenced key is either primary or unique and contains
all columns as in foreign key in same order.
3. Added new test cases to existing MTR test main.foreign_key.test
4. Added new MTR test fk_on_std_key_upgrade.test to check upgrade
from 8.x

Change-Id: I0df7eab36c548fc282570ade3e03c085b884c181