Bug#36159317 INNODB: DEADLOCK BETWEEN FIL_SHARD::OPEN_FILE, FIL_OP_REPLAY_RENAME ON WINDOWS 2
Symptom:

Tests involving RENAME deadlock on Windows

Root cause:

The original fix for the
Bug#32808809 INNODB: DEADLOCK BETWEEN FIL_SHARD::OPEN_FILE, FIL_OP_REPLAY_RENAME ON WINDOWS
missed to call to guard.rollback() before fil_space_read_name_and_filepath(),
which would be needed to actually close() the file handle.
The fil_space_read_name_and_filepath() tries to acquire Fil_shard mutex,
which risks a deadlock if done while keeping the file open, because
usual ordering of these two operations is the opposite: first latch the
Fil_shard mutex, and only then open the file.
In particular page cleaners writing dirty pages to tablespace files do
it in this usual order, and thus can deadlock with a RENAME which does
it in the opposite order.

Fix:

Turns out we don't need a guard at all, because we can simply close the
file handle right after a call to read_first_page().

Also, when trying to prove correctness of the patch, it turned out that
there is some confusion in the code about the lifecycle of the file handle.
The doxygen for read_first_page() says the file handle should be
open before calling it, which would be a useful property to reason about
the code, if really enforced.
Indeed most callers ensured this, yet the read_first_page() had a
defensive conditional logic to open the file handle if it is closed,
just in case.
Turned out that there was one place in the code where we tried to call
validate_first_page() while the file handle was closed, and that place
was in itself buggy, in SysTablespace::read_lsn_and_check_flags():

   for (int retry = 0; retry < 2; ++retry) {
     err = it->validate_first_page(it->m_space_id, flushed_lsn, false);

     if (err != DB_SUCCESS &&
         (retry == 1 || it->open_or_create(srv_read_only_mode) != DB_SUCCESS ||
          it->restore_from_doublewrite(0) != DB_SUCCESS)) {
       it->close();

       return (err);
     }
   }

clearly the intent here was to retry in case of error, yet the buggy
logic was retrying even if there was no error. Because the contract of
validate_first_page() specified in doxygen clearly says the file handle
is closed after the call, the second call occured while handle was closed.
This patch fixes the problem by unrolling this loop, so that it is easier
to understand and verify its logic.
This made it possible for this patch to, also remove the conditional
opening of the file handle from read_first_page() to match its contract.

The rewrite of the loop also made it clear that there was another bug
in a code which was not covered with tests:

   /* Make sure the tablespace space ID matches the
   space ID on the first page of the first datafile. */
   if (space_id() != it->m_space_id) {
     ib::error(ER_IB_MSG_444)
         << "The " << name() << " data file '" << it->name()
         << "' has the wrong space ID. It should be " << space_id() << ", but "
         << it->m_space_id << " was found";

     it->close();

     return (err);
   }

The err was provably always DB_SUCCESS, so the relatively serious error
of not having the right space_id, was ignored by the caller.
Turns out that this code is not really needed, because the
validate_first_page() we called earlier can help us verify the space_id.
However there was yet another bug here: instead of passing the expected
space_id() to it, we were passing the it->m_space_id found on disc -
which defeated the purpose of validation, because it was comparing the
value read from disc to itself.
This patch fixes it by removing the buggy error handling and passing
space_id() instead of it->m_space_id to validate_first_page.

Change-Id: I05230d01e5bfc39fe97f645100a8db212c1219c8