Bug#36004840 redo log consumer cannot advance
Advancing the redo log cursor created with
    DO innodb_redo_log_consumer_register()
was requiring a BACKUP_ADMIN priviledge, and the check for this
priviledge could be blocked by concurrent USER-related operations such
as CREATE USER. And CREATE USER attempts writing data to redo log, which
in turn is blocked if there's not enough space in redo log until the old
redo logs are reclaimed. And redo logs can't be reclaimed as long as
there is a redo log consumer whose cursor was not advanced past them.
This caused a deadlock, which could be resolved even by increasing
redo log capacity in theory, but in practice in order to issue such
command one needs a client connected to the server, but esablishing new
sessions was also blocked by such deadlock.

The solution is to not check for BACKUP_ADMIN priviledge during
    DO innodb_redo_log_consumer_advance(x)
as the operation is safe: the only effect of it is that it could unblock
the reclamation of old redo logs kept around for the current's session
consumer, which is an operation which can help the server, not bring it
down. What matters is that we check for BACKUP_ADMIN priviledge when the
    DO innodb_redo_log_consumer_register()
is called to establish the cursor - which *is* a risky operation
precisely because of the possibility to cause deadlocks such as this one.
If the user's session somehow lost the priviledge meanwhile, it is still
better to let it advance the cursor then to try to refuse it.

Change-Id: I0019aca8b2528f02864146acf9cc7f61d1c3522d