Bug#35529604 : Connect from OCI instance with PHP 5.6 to DBS                (mysql 8.0.33) get the following error:                "PHP message: PHP Warning: mysql_connect():                Server sent charset (255) unknown to the                client"
Description:
------------
When the character_set_server variable is set using
SET PERSIST or SET GLOBAL, it does not take effect for the
new client sessions or for a client that is establishing a
connection to the server after the server is restarted. The
only way currently we can achieve this is by setting this
variable via command line when the server is started.

Fix:
----
The fix is to make sure that at the time of server restart,
the data is read in the correct order so that the variable
takes effect as expected.

Change-Id: I8fb468efaa4492d00549dc47b975b4fdfe0ec970