Bug#36151537: Assertion `m_typelib != nullptr' failed [1/2, NULLIF]
Calls to NULLIF with an ENUM or SET value as first argument could
cause assertion failures during resolving.

The problem was that Item_func_nullif didn't implement get_typelib()
that provides metadata for the ENUM or SET type.

Fixed by implementing the missing member function. Since the return
type of NULLIF is the same as the first argument,
Item_func_nullif::get_typelib() delegates the call to the first
argument.

Change-Id: Ie74aa37598ec26cf07763b7b2ce8be6c192e810c