Bug#36047803 IMPORT TABLESPACE not friendly with multi-byte characters
This patch is based on contribution from Lee Adria.
Thank you for contributing to MySQL.

Problem
-------
User is unable to import the tablespace that has long column name with
multi-byte characters.

Description
-----------
At the time of importing the tablespace, innodb validates the column
length that is read from the meta data file(.cfg) against a fixed length
of 128 bytes, if length is more than 128 bytes then innodb throws error.
However server supports the max column name length of 64 chars including
multi-byte chars.  Thus the assumption in the current code is wrong.
The patch proposed by Lee validates the length with allowed maximum
length of column that is 64 multi-byte chars.

Change-Id: I7bdf04a9f58bb8c1529e6f95cf20a34438584498