Bug#36151537: Assertion `m_typelib != nullptr' failed [2/2, IF/CASE/COALESCE]
Fix for some related problems.

Some implementations of Item::get_typelib() assume that the function
is not called on items that are not of ENUM or SET type. This
assumption does not hold, as some other implementations call
get_typelib() on its arguments in order to check whether or not their
type is ENUM or SET. This inconsistency leads to assertion failures.

Fixed by changing the affected functions so that they return nullptr,
without raising assertion failures, if the item is not of ENUM or SET
type.

Change-Id: I505c4a9e901b32ff07eee4c16e62fd1f6cc9e2ff