Bug#36164105 New section for 8.2 error messages added in share/messages_to_error_log
The patch for WL#15786 added a new error log message and placed it in a
new section in the messages_to_error_log.txt file. This went against the
convention that all 8.x errors should be placed in the same section (a
new section should not be added before 9.0). This patch removes the new
section and moves the error to the end of the 8.x section.

Change-Id: Ia0d76e973a2a4c037b3d144fb9084ff8eb033980