Bug#35169384: Result diff seen with hypergraph off and on (different number of rows)
Despite the bug title, this bug affected both the old optimizer and
the hypergraph optimizer. Different results were seen in a query
depending on whether an index range scan was chosen or not.

The bug was caused by incorrect handling of non-binary strings in the
range analysis when a string constant was longer than the indexed
column. It assumed that if the constant string in the range predicate
was longer than the data type of the indexed column, it was OK to use
the prefix of the constant as range boundary instead, possibly
adjusting < to <= to not miss any qualifying row, and using a filter
to remove the extra rows that may be returned by the range scan on the
prefix.

This is fine when the string has a binary collation. With non-binary
collations, it is more complicated. The constant string could be
truncated in the middle of a sequence that the collation handles as
equal to some shorter sequence of characters. Or the prefix of the
constant string could contain characters that are ignored in the
collation (like SOFT HYPHEN). So it's not safe to say that it's enough
to compare the MIN(N, M) first characters of a VARCHAR(N) and a
VARCHAR(M) to find out if it's a potential match.

Examples of problems if we have an index on a VARCHAR(1) column with
the default utf8mb4_0900_ai_ci collation:

'ß' = 'ss', so a predicate such as c <= 'ss' should match the value
'ß'. But when the range scan uses c <= 's', it misses that value,
because 's' < 'ss' = 'ß'.

SOFT HYPHEN (U+00AD) is ignored, so if you have a predicate that
compares the indexed column to a constant that starts with a SOFT
HYPHEN, say c <= CONCAT(_utf16 0x00ad, 'a'), and this is translated to
the range predicate c <= SOFT HYPEN, the range predicate is equivalent
to c <= '', so the index range scan will not find the value 'a', which
both fits in the indexed column and matches the original predicate.

It is fixed by refusing to create an index range predicate from a
predicate that compares string column to a constant string that is
longer than the maximum length of the column, unless the collation of
the column has some properties that makes it safe. An ordinary filter
will be used instead for these predicates.

Safe collations are binary collations, as well as those that produce
weight strings only via the CHARSET_INFO::sort_order table (mainly
some of the legacy collations for single-byte character sets, like
latin1_swedish_ci).

Some tests with recorded query plans failed because of plan changes.
These tests have been updated either by recording the new plan as the
expected one, or by slightly changing the test cases so that they
still can use an index range plan.

Change-Id: I5eab7456c1cd00fefe841d5023bcfe7a71d93518