Bug#35781732 5.7 vs 8.0 user variables differences output in  performance_schema.user_variables_by_thread
Problem:
The value stored in performance_schema.user_variables_by_thread
table becomes an INT with a rounded value of the user variable.
Value that is being stored in the table is integer and not
decimal in the variable_value field.

Fix:
The performance schema code for user_variables_by_thread table
is explicitly assigning the decimal precision to be zero,
which is causing the performance_schema.user_variables_by_thread
table to store variable_value field in rounded off integer.
The fix is to change deciaml precision to DECIMAL_NOT_SPECIFIED.
This will prevent rounding off the variable_value
field and print the out decimal value instead of integer.

Change-Id: I76074be0097b58fd872e332235f13ed05a8a68f9