Bug#35789759 - omit multi-value index columns when extracting rows from fetch cache
Symptom: When using a multi-valued index, the multi-valued fields
are incorrectly copied on some SELECT queries leading to
unpredictable results.

Root cause: When retrieving rows from fetch cache, virtual
columns are skipped unless an index-only scan is performed on an
index which is based on these virtual columns. This is because
the column values may be needed, but impossible to calculate by
the server. Multi-values are an exception to this - they are
never needed by the server and cannot be actually copied, because
the result is reduced from multiple index rows containing the
different individual values into a single result per table row.
The code handling the copy did not account for multi-value
and attempted to copy them; the actual buffer contents however
do not in this case conform to the binary representation of the
column type.

Solution: When copying rows from fetch buffer field-by-field,
always skip multi-value columns.

Change-Id: I4b950872ec4de1bd73ef5f3bfd61b019a71eb29f