Bug#36172977 create_clusterset() and set_mock_clusterset_metadata() API is hard to expand
create_clusterset() and set_mock_clusterset_metadata()  are the functions used
in the component tests of the Router ClusterSet handling.
Their API is cumbersome with many default parameters...
The caller has to repeat them even if default if it wants to set some specific
that is located further on the parmeters list.
Adding a new parameter is also a challenge as they go at the end and at some
point their order is completely random and messy.

This patch changes the API to accept an object with fields initialized with
default values. The caller is supposed to set the non-default values and pass
the object as a function parameter.

Change-Id: Ie7623ade862126389ee1422af37995cd584fba17