Bug#35784192 tablespace is missing for table innodb_undo_002
This bug reports following two problems:

Querying the information schema reports undo tablespace is missing -
  while an undo file is present.
Purge threads don't remove the undo_space_number_trunc.log file.

While the first problem is already fixed through Bug#32104924, Bug#32654667

This bug fix aims to address the second problem.

The purge thread attempts to lock dictionary tables, while truncating
the undo tablespace. This lock can't be acquired in
a read-only MySQL instance, except when the thread has some special
flags associated with it. In this case, we use the skip_readonly_check
member of the thread descriptor struct, to signal that the readonly
check should not be applied to the purge thread. This is done
via the set_skip_readonly_check() method.

Change-Id: Iea6255ff6654752db8658bdd422cc3e58e3a8a97