Bug#36113505 Revive NdbMutex_DeadlockDetector
Patch is a fix of the deadlock *prediction* implemented by comparing the
lock order we have previously seen among the mutex and collected in the
before/after set of mutexes.

Existing implementation detected lots of false positives as it did not
take the different combinations of the locks seen into consideration.
- It just flagged a deadlock if we now try to take a lock which has
previously been taken 'before' any of the other lock we already hold.

That turned out to be a too course 'filter' to decide deadlocking combinations.
Thus this patch introduce collection of all the different lock combinations
held 'before' each lock and check these when a new lock is to be taken.

We also needed to rule out lock combinations where a common set of lock
where held 'before', both by the current thread and the other locks
combination seen before, e.g, if

  before locks ---- thread now hold  --- try to take
  [x, A, B]           [x, B]               [ A ]

The reverse lock order of A,B, vs now trying to lock B,A could have
been a classic deadlock. However as the before-combination also held
the x-lock, which this thread also hold, it prevent the A/B locks
from being taken at the same time -> no deadlocks

Patch also add excemption logic for 'non-blocking' locks, that is:
  - Try-locks
  - 'Locks' not implemented as a mutex, but use some 'state' info
    to guard against multiple thread entering  the same code region/state

The later is supported by the new functions NdbMutex_EnterSerializedRegion()
and NdbMutex_LeaveSerializedRegion() which enter the 'lock' information into
the deadlock detector internal structures.

Patch also has some refactoring of the DD-data structures
1) Moved struct decl from header file into cpp-file
2) Utilty functions made struct member methods.
3) Removed 'extern C' declaration, it is now all C++
4) Changed some arguments from taking a NdbMutex* to a mutex_state* instead
5) Renamed variables from 'm', 't' and such to take fuller names
6) Used std::vector to implement the arrays.

Note that the patch diff might appear larger than it actually is due to
the above.

Change-Id: Ibe60ff4f82b35d21a12f6e016c834ff54197d92d