Bug#36176983 Improve binlog purge logging
Problem:
It's hard to diagnose error related to purging of rows from
mysql.ndb_binlog_index. There seem to be cases when the purge is slow or
locks up. Slow purging may affect binlog injector.

Analysis:
To know when purgin start and end is vital for being able to know
duration of purge and wheter or not other parts of system is affected by
purge.

Solution:
Improve logging related to when purging starts and completes.

After this the log file will have these kind of printouts:

- When binlog injector recall pending purges:
  <time> 2 [System] [MY-010866] [NDB] Binlog: Purging binlog file: './
binlog.000078'
  <time> 2 [System] [MY-010866] [NDB] Binlog: Purged binlog file: './
binlog.000078'

- When purging binlog:
  <time> 0 [System] [MY-010865] [NDB] Purging binlog file: './binlog.
000001'
  <time> 0 [System] [MY-010865] [NDB] Purged binlog file: './binlog.
000001'

Change-Id: I457fc347e7f95f5dc9981dd7d42dfff6385681f5